refs #6, #27: Refactor memory pools and host/device memory types
 * Tried to support OpenCL, but confirmed that we need OpenCL 2.0+.
   (which is not supported by current generation of Xeon Phi...)

   - Related codes will be rewritten someday using SVM (shared virtual
     memory) APIs in OpenCL 2.0+.

 * Reduced memory footprint of batch_ids array passed to the device.

 * Rollbacked ev_prepare -> ev_check watcher type change (c732a25),
   as it has broken CPU-only cross-node IP forwarding scenarios. :(

 * TODO: fix IPsec GPU-only mode...

+++ elements/ip/IPlookup_kernel.cu
@@ -33,7 +33,7 @@ __device__ uint32_t ntohl(uint32_t n)
 /* The GPU kernel. */
 __global__ void ipv4_route_lookup_cuda(
         struct datablock_kernel_arg **datablocks,
-        uint32_t count, uint16_t *batch_ids, uint16_t *item_ids,
+        uint32_t count, uint8_t *batch_ids, uint16_t *item_ids,
         uint8_t *checkbits_d,
         uint16_t* __restrict__ TBL24_d,
         uint16_t* __restrict__ TBLlong_d)
